added support for "sanitize_callback" functions
In order to encourage and allow users to validate and sanitize data
before sending it to the database, "sanitize_callback" function support
has been built into the class. As part of the "fields" $opts subarray, a
user can define a "sanitize_callback" function to be called prior to the
data being written to the database. The "sanitize_callback" function
should return the value that will be saved to the database.